Was successfully able to compile and run C++ wrapped functions for both PyConnections and PyInteractions.
