--- cpp_src/cython/network.cpp
@@ -1,1 +1,0 @@
-#error Do not use this file, it is the result of a failed Cython compilation.
