+++ cpp_src/cython/simfection_cpp.pxd
@@ -1,11 +1,27 @@
 from libcpp.vector cimport vector
-from libcpp.map cimport map
-from libcpp.utility cimport pair
 from libcpp cimport bool
-from libcpp.string import string
+from libcpp.unordered_map cimport unordered_map
+from libcpp.utility cimport pair
+from libcpp.string cimport string
+
+# Declare the class with cdef
+cdef extern from "simfection_cpp.h" namespace "Connections":
+    cdef cppclass Connections:
+        # private:
+        int size 
+        # public: 
+        Connections() except +
+        Connections(int) except +
+        int getSize()
+        void setSize(int)
+        bool existsInVector(int, vector[int])
+        vector[int] genConnectionsMaxVector(int, int, int)
+        vector[int] getAvailable(int, vector[int], vector[vector[int]])
+        vector[vector[int]] genRandomNetwork(vector[int], bool, bool)
+
 
 # Declare the Population nested/inner class
-cdef extern from "interactions.h" namespace "Interactions::Population":
+cdef extern from "simfection_cpp.h" namespace "Interactions::Interactions::Population":
     cdef cppclass Population:
     # private:
         vector[int] agents
@@ -13,7 +29,7 @@ cdef extern from "interactions.h" namespace "Interactions::Population":
         vector[vector[int]] infectedByList
         vector[int] daysInfected
         vector[int] immunities
-        map[int, string] agentStates
+        unordered_map[int, string] agentStates
         # public:
         Population() except +
         Population(vector[int],
@@ -27,7 +43,7 @@ cdef extern from "interactions.h" namespace "Interactions::Population":
         vector[vector[int]] getInfectedByList()
         vector[int] getDaysInfected()
         vector[int] getImmunities()
-        map[int, string] getAgentStates()
+        unordered_map[int, string] getAgentStates()
         # Setters
         void setAgents(vector[int])
         void setStates(vector[string])
@@ -38,17 +54,17 @@ cdef extern from "interactions.h" namespace "Interactions::Population":
 
 
 # Declare the Connections nested/inner class
-cdef extern from "interactions.h" namespace "Interactions::Connections":
+cdef extern from "simfection_cpp.h" namespace "Interactions::Interactions::InteractConnections":
     # Connections subclass
-    cdef cppclass Connections:
+    cdef cppclass InteractConnections:
         # private:
         vector[int] agents
         vector[vector[int]] connectionsList
         vector[int] numConnections
         vector[int] maxConnections
         # public: 
-        Connections() except +
-        Connections(vector[int], 
+        InteractConnections() except +
+        InteractConnections(vector[int], 
                     vector[vector[int]], 
                     vector[int],
                     vector[int]) except +
@@ -65,7 +81,7 @@ cdef extern from "interactions.h" namespace "Interactions::Connections":
 
 
 # Declare the Interactions class
-cdef extern from "interactions.h" namespace "Interactions":
+cdef extern from "simfection_cpp.h" namespace "Interactions":
     cdef cppclass Interactions:
         # public: 
         # Constructor
@@ -80,23 +96,23 @@ cdef extern from "interactions.h" namespace "Interactions":
                             vector[vector[int]], 
                             vector[int],
                             vector[int])
-        void setPathogenSettings(map[string, float])
+        void setPathogenSettings(unordered_map[string, float])
         # Getters
         Population getPopulation()
-        Connections getConnections()
-        map[string, float] getPathogenSettings()
+        InteractConnections getConnections()
+        unordered_map[string, float] getPathogenSettings()
         pair[int, string] getAgentIDAndStatePair(int)
         # Interaction Engine Functions
         bool isSameString(string str1, string str2)
-        vector[vector[int]] getUniqueConnections(Connections thisConnections)
+        vector[vector[int]] getUniqueConnections(InteractConnections thisConnections)
         pair[bool, pair[int, int]] qualify_interaction(pair[pair[int, string], pair[int, string]])
         void interact(vector[int])
         void interactAll()
         # private:
         Population population
-        Connections connections
+        InteractConnections connections
         string inf
         string sus
         string dead
-        map[string, float] pathogenSettings
+        unordered_map[string, float] pathogenSettings
